+++ index.js
@@ -36,9 +36,10 @@ const renderPermalink = (slug, opts, state, idx) => {
 const uniqueSlug = (slug, slugs, failOnNonUnique) => {
   let uniq = slug
   let i = 2
-  while (hasProp.call(slugs, uniq)) uniq = `${slug}-${i++}`
-  if (uniq !== slug && failOnNonUnique) {
-    throw new Error(`Defined slug/ID '${slug}' is not unique. Please fix this ID duplication.`);
+  if (failOnNonUnique && hasProp.call(slugs, uniq)) {
+    throw Error(`User defined id attribute '${slug}' is NOT unique. Please fix it in your markdown to continue.`)
+  } else {
+    while (hasProp.call(slugs, uniq)) uniq = `${slug}-${i++}`
   }
   slugs[uniq] = true
   return uniq
@@ -71,13 +72,11 @@ const anchor = (md, opts) => {
         let slug = token.attrGet('id')
 
         if (slug == null) {
-          slug = uniqueSlug(opts.slugify(title), slugs, false);
+          slug = uniqueSlug(opts.slugify(title), slugs, false)
         } else {
-          // mark existing slug/ID as unique, at least.
-          // IFF it collides, FAIL!
-          slug = uniqueSlug(slug, slugs, true);
+          slug = uniqueSlug(slug, slugs, true)
         }
-        token.attrSet('id', slug);
+        token.attrSet('id', slug)
 
         if (opts.permalink) {
           opts.renderPermalink(slug, opts, state, tokens.indexOf(token))